use kv_transfer_infinistore
